P-chain validator management refactor (#1284)
