+++ vms/platformvm/blocks/executor/acceptor.go
@@ -9,13 +9,12 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/snow/choices"
 	"github.com/ava-labs/avalanchego/utils"
-	"github.com/ava-labs/avalanchego/utils/window"
 	"github.com/ava-labs/avalanchego/vms/platformvm/blocks"
 	"github.com/ava-labs/avalanchego/vms/platformvm/metrics"
 	"github.com/ava-labs/avalanchego/vms/platformvm/state"
+	"github.com/ava-labs/avalanchego/vms/platformvm/validators"
 )
 
 var (
@@ -29,9 +28,9 @@ var (
 // being shutdown.
 type acceptor struct {
 	*backend
-	metrics          metrics.Metrics
-	recentlyAccepted window.Window[ids.ID]
-	bootstrapped     *utils.Atomic[bool]
+	metrics      metrics.Metrics
+	validators   validators.Manager
+	bootstrapped *utils.Atomic[bool]
 }
 
 func (a *acceptor) BanffAbortBlock(b *blocks.BanffAbortBlock) error {
@@ -316,6 +315,6 @@ func (a *acceptor) commonAccept(b blocks.Block) error {
 	a.state.SetLastAccepted(blkID)
 	a.state.SetHeight(b.Height())
 	a.state.AddStatelessBlock(b, choices.Accepted)
-	a.recentlyAccepted.Add(blkID)
+	a.validators.OnAcceptedBlockID(blkID)
 	return nil
 }
