+++ vms/platformvm/blocks/executor/acceptor_test.go
@@ -5,7 +5,6 @@ package executor
 
 import (
 	"testing"
-	"time"
 
 	"github.com/golang/mock/gomock"
 
@@ -19,12 +18,12 @@ import (
 	"github.com/ava-labs/avalanchego/utils"
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/utils/timer/mockable"
-	"github.com/ava-labs/avalanchego/utils/window"
 	"github.com/ava-labs/avalanchego/vms/components/verify"
 	"github.com/ava-labs/avalanchego/vms/platformvm/blocks"
 	"github.com/ava-labs/avalanchego/vms/platformvm/metrics"
 	"github.com/ava-labs/avalanchego/vms/platformvm/state"
 	"github.com/ava-labs/avalanchego/vms/platformvm/txs"
+	"github.com/ava-labs/avalanchego/vms/platformvm/validators"
 	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 )
 
@@ -61,8 +60,8 @@ func TestAcceptorVisitProposalBlock(t *testing.T) {
 			},
 			state: s,
 		},
-		metrics:          metrics.Noop,
-		recentlyAccepted: nil,
+		metrics:    metrics.Noop,
+		validators: validators.TestManager,
 	}
 
 	err = acceptor.ApricotProposalBlock(blk)
@@ -98,12 +97,8 @@ func TestAcceptorVisitAtomicBlock(t *testing.T) {
 				SharedMemory: sharedMemory,
 			},
 		},
-		metrics: metrics.Noop,
-		recentlyAccepted: window.New[ids.ID](window.Config{
-			Clock:   &mockable.Clock{},
-			MaxSize: 1,
-			TTL:     time.Hour,
-		}),
+		metrics:    metrics.Noop,
+		validators: validators.TestManager,
 	}
 
 	blk, err := blocks.NewApricotAtomicBlock(
@@ -183,12 +178,8 @@ func TestAcceptorVisitStandardBlock(t *testing.T) {
 				SharedMemory: sharedMemory,
 			},
 		},
-		metrics: metrics.Noop,
-		recentlyAccepted: window.New[ids.ID](window.Config{
-			Clock:   clk,
-			MaxSize: 1,
-			TTL:     time.Hour,
-		}),
+		metrics:    metrics.Noop,
+		validators: validators.TestManager,
 	}
 
 	blk, err := blocks.NewBanffStandardBlock(
@@ -278,12 +269,8 @@ func TestAcceptorVisitCommitBlock(t *testing.T) {
 				SharedMemory: sharedMemory,
 			},
 		},
-		metrics: metrics.Noop,
-		recentlyAccepted: window.New[ids.ID](window.Config{
-			Clock:   &mockable.Clock{},
-			MaxSize: 1,
-			TTL:     time.Hour,
-		}),
+		metrics:      metrics.Noop,
+		validators:   validators.TestManager,
 		bootstrapped: &utils.Atomic[bool]{},
 	}
 
@@ -374,12 +361,8 @@ func TestAcceptorVisitAbortBlock(t *testing.T) {
 				SharedMemory: sharedMemory,
 			},
 		},
-		metrics: metrics.Noop,
-		recentlyAccepted: window.New[ids.ID](window.Config{
-			Clock:   &mockable.Clock{},
-			MaxSize: 1,
-			TTL:     time.Hour,
-		}),
+		metrics:      metrics.Noop,
+		validators:   validators.TestManager,
 		bootstrapped: &utils.Atomic[bool]{},
 	}
