+++ vms/platformvm/vm.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"time"
 
 	"github.com/gorilla/rpc/v2"
 
@@ -18,7 +17,6 @@ import (
 	"github.com/ava-labs/avalanchego/cache"
 	"github.com/ava-labs/avalanchego/codec"
 	"github.com/ava-labs/avalanchego/codec/linearcodec"
-	"github.com/ava-labs/avalanchego/database"
 	"github.com/ava-labs/avalanchego/database/manager"
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/snow"
@@ -33,7 +31,6 @@ import (
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/utils/math"
 	"github.com/ava-labs/avalanchego/utils/timer/mockable"
-	"github.com/ava-labs/avalanchego/utils/window"
 	"github.com/ava-labs/avalanchego/utils/wrappers"
 	"github.com/ava-labs/avalanchego/version"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
@@ -53,12 +50,7 @@ import (
 	blockexecutor "github.com/ava-labs/avalanchego/vms/platformvm/blocks/executor"
 	txbuilder "github.com/ava-labs/avalanchego/vms/platformvm/txs/builder"
 	txexecutor "github.com/ava-labs/avalanchego/vms/platformvm/txs/executor"
-)
-
-const (
-	validatorSetsCacheSize        = 512
-	maxRecentlyAcceptedWindowSize = 256
-	recentlyAcceptedWindowTTL     = 5 * time.Minute
+	pvalidators "github.com/ava-labs/avalanchego/vms/platformvm/validators"
 )
 
 var (
@@ -68,12 +60,12 @@ var (
 	_ validators.SubnetConnector = (*VM)(nil)
 
 	errMissingValidatorSet = errors.New("missing validator set")
-	errMissingValidator    = errors.New("missing validator")
 )
 
 type VM struct {
 	config.Config
 	blockbuilder.Builder
+	validators.State
 
 	metrics            metrics.Metrics
 	atomicUtxosManager avax.AtomicUTXOManager
@@ -94,14 +86,6 @@ type VM struct {
 
 	// Bootstrapped remembers if this chain has finished bootstrapping or not
 	bootstrapped utils.Atomic[bool]
-
-	// Maps caches for each subnet that is currently tracked.
-	// Key: Subnet ID
-	// Value: cache mapping height -> validator set map
-	validatorSetCaches map[ids.ID]cache.Cacher[uint64, map[ids.NodeID]*validators.GetValidatorOutput]
-
-	// sliding window of blocks that were recently accepted
-	recentlyAccepted window.Window[ids.ID]
 
 	txBuilder txbuilder.Builder
 	manager   blockexecutor.Manager
@@ -143,15 +127,6 @@ func (vm *VM) Initialize(
 		return err
 	}
 
-	vm.validatorSetCaches = make(map[ids.ID]cache.Cacher[uint64, map[ids.NodeID]*validators.GetValidatorOutput])
-	vm.recentlyAccepted = window.New[ids.ID](
-		window.Config{
-			Clock:   &vm.clock,
-			MaxSize: maxRecentlyAcceptedWindowSize,
-			TTL:     recentlyAcceptedWindowTTL,
-		},
-	)
-
 	rewards := reward.NewCalculator(vm.RewardConfig)
 	vm.state, err = state.New(
 		vm.dbManager.Current().Database,
@@ -167,6 +142,8 @@ func (vm *VM) Initialize(
 		return err
 	}
 
+	validatorManager := pvalidators.NewManager(vm.Config, vm.state, vm.metrics, &vm.clock)
+	vm.State = validatorManager
 	vm.atomicUtxosManager = avax.NewAtomicUTXOManager(chainCtx.SharedMemory, txs.Codec)
 	utxoHandler := utxo.NewHandler(vm.ctx, &vm.clock, vm.fx)
 	vm.uptimeManager = uptime.NewManager(vm.state)
@@ -205,7 +182,7 @@ func (vm *VM) Initialize(
 		vm.metrics,
 		vm.state,
 		txExecutorBackend,
-		vm.recentlyAccepted,
+		validatorManager,
 	)
 	vm.Builder = blockbuilder.New(
 		mempool,
@@ -290,18 +267,18 @@ func (vm *VM) onNormalOperationsStarted() error {
 		return err
 	}
 
-	primaryVdrIDs, exists := vm.getValidatorIDs(constants.PrimaryNetworkID)
-	if !exists {
-		return errMissingValidatorSet
+	primaryVdrIDs, err := validators.NodeIDs(vm.Validators, constants.PrimaryNetworkID)
+	if err != nil {
+		return err
 	}
 	if err := vm.uptimeManager.StartTracking(primaryVdrIDs, constants.PrimaryNetworkID); err != nil {
 		return err
 	}
 
 	for subnetID := range vm.TrackedSubnets {
-		vdrIDs, exists := vm.getValidatorIDs(subnetID)
-		if !exists {
-			return errMissingValidatorSet
+		vdrIDs, err := validators.NodeIDs(vm.Validators, subnetID)
+		if err != nil {
+			return err
 		}
 		if err := vm.uptimeManager.StartTracking(vdrIDs, subnetID); err != nil {
 			return err
@@ -337,18 +314,18 @@ func (vm *VM) Shutdown(context.Context) error {
 	vm.Builder.Shutdown()
 
 	if vm.bootstrapped.Get() {
-		primaryVdrIDs, exists := vm.getValidatorIDs(constants.PrimaryNetworkID)
-		if !exists {
-			return errMissingValidatorSet
+		primaryVdrIDs, err := validators.NodeIDs(vm.Validators, constants.PrimaryNetworkID)
+		if err != nil {
+			return err
 		}
 		if err := vm.uptimeManager.StopTracking(primaryVdrIDs, constants.PrimaryNetworkID); err != nil {
 			return err
 		}
 
 		for subnetID := range vm.TrackedSubnets {
-			vdrIDs, exists := vm.getValidatorIDs(subnetID)
-			if !exists {
-				return errMissingValidatorSet
+			vdrIDs, err := validators.NodeIDs(vm.Validators, subnetID)
+			if err != nil {
+				return err
 			}
 			if err := vm.uptimeManager.StopTracking(vdrIDs, subnetID); err != nil {
 				return err
@@ -366,21 +343,6 @@ func (vm *VM) Shutdown(context.Context) error {
 		vm.dbManager.Close(),
 	)
 	return errs.Err
-}
-
-func (vm *VM) getValidatorIDs(subnetID ids.ID) ([]ids.NodeID, bool) {
-	validatorSet, exist := vm.Validators.Get(subnetID)
-	if !exist {
-		return nil, false
-	}
-	validators := validatorSet.List()
-
-	validatorIDs := make([]ids.NodeID, len(validators))
-	for i, vdr := range validators {
-		validatorIDs[i] = vdr.NodeID
-	}
-
-	return validatorIDs, true
 }
 
 func (vm *VM) ParseBlock(_ context.Context, b []byte) (snowman.Block, error) {
@@ -473,200 +435,6 @@ func (vm *VM) Disconnected(_ context.Context, nodeID ids.NodeID) error {
 		return err
 	}
 	return vm.state.Commit()
-}
-
-// GetValidatorSet returns the validator set at the specified height for the
-// provided subnetID.
-func (vm *VM) GetValidatorSet(ctx context.Context, height uint64, subnetID ids.ID) (map[ids.NodeID]*validators.GetValidatorOutput, error) {
-	validatorSetsCache, exists := vm.validatorSetCaches[subnetID]
-	if !exists {
-		validatorSetsCache = &cache.LRU[uint64, map[ids.NodeID]*validators.GetValidatorOutput]{Size: validatorSetsCacheSize}
-		// Only cache tracked subnets
-		if subnetID == constants.PrimaryNetworkID || vm.TrackedSubnets.Contains(subnetID) {
-			vm.validatorSetCaches[subnetID] = validatorSetsCache
-		}
-	}
-
-	if validatorSet, ok := validatorSetsCache.Get(height); ok {
-		vm.metrics.IncValidatorSetsCached()
-		return validatorSet, nil
-	}
-
-	lastAcceptedHeight, err := vm.GetCurrentHeight(ctx)
-	if err != nil {
-		return nil, err
-	}
-	if lastAcceptedHeight < height {
-		return nil, database.ErrNotFound
-	}
-
-	// get the start time to track metrics
-	startTime := vm.Clock().Time()
-
-	currentSubnetValidators, ok := vm.Validators.Get(subnetID)
-	if !ok {
-		currentSubnetValidators = validators.NewSet()
-		if err := vm.state.ValidatorSet(subnetID, currentSubnetValidators); err != nil {
-			return nil, err
-		}
-	}
-	currentPrimaryNetworkValidators, ok := vm.Validators.Get(constants.PrimaryNetworkID)
-	if !ok {
-		// This should never happen
-		return nil, errMissingValidatorSet
-	}
-
-	currentSubnetValidatorList := currentSubnetValidators.List()
-	vdrSet := make(map[ids.NodeID]*validators.GetValidatorOutput, len(currentSubnetValidatorList))
-	for _, vdr := range currentSubnetValidatorList {
-		primaryVdr, ok := currentPrimaryNetworkValidators.Get(vdr.NodeID)
-		if !ok {
-			// This should never happen
-			return nil, fmt.Errorf("%w: %s", errMissingValidator, vdr.NodeID)
-		}
-		vdrSet[vdr.NodeID] = &validators.GetValidatorOutput{
-			NodeID:    vdr.NodeID,
-			PublicKey: primaryVdr.PublicKey,
-			Weight:    vdr.Weight,
-		}
-	}
-
-	for i := lastAcceptedHeight; i > height; i-- {
-		weightDiffs, err := vm.state.GetValidatorWeightDiffs(i, subnetID)
-		if err != nil {
-			return nil, err
-		}
-
-		for nodeID, weightDiff := range weightDiffs {
-			vdr, ok := vdrSet[nodeID]
-			if !ok {
-				// This node isn't in the current validator set.
-				vdr = &validators.GetValidatorOutput{
-					NodeID: nodeID,
-				}
-				vdrSet[nodeID] = vdr
-			}
-
-			// The weight of this node changed at this block.
-			var op func(uint64, uint64) (uint64, error)
-			if weightDiff.Decrease {
-				// The validator's weight was decreased at this block, so in the
-				// prior block it was higher.
-				op = math.Add64
-			} else {
-				// The validator's weight was increased at this block, so in the
-				// prior block it was lower.
-				op = math.Sub[uint64]
-			}
-
-			// Apply the weight change.
-			vdr.Weight, err = op(vdr.Weight, weightDiff.Amount)
-			if err != nil {
-				return nil, err
-			}
-
-			if vdr.Weight == 0 {
-				// The validator's weight was 0 before this block so
-				// they weren't in the validator set.
-				delete(vdrSet, nodeID)
-			}
-		}
-
-		pkDiffs, err := vm.state.GetValidatorPublicKeyDiffs(i)
-		if err != nil {
-			return nil, err
-		}
-
-		for nodeID, pk := range pkDiffs {
-			// pkDiffs includes all primary network key diffs, if we are
-			// fetching a subnet's validator set, we should ignore non-subnet
-			// validators.
-			if vdr, ok := vdrSet[nodeID]; ok {
-				// The validator's public key was removed at this block, so it
-				// was in the validator set before.
-				vdr.PublicKey = pk
-			}
-		}
-	}
-
-	// cache the validator set
-	validatorSetsCache.Put(height, vdrSet)
-
-	endTime := vm.Clock().Time()
-	vm.metrics.IncValidatorSetsCreated()
-	vm.metrics.AddValidatorSetsDuration(endTime.Sub(startTime))
-	vm.metrics.AddValidatorSetsHeightDiff(lastAcceptedHeight - height)
-	return vdrSet, nil
-}
-
-// GetCurrentHeight returns the height of the last accepted block
-func (vm *VM) GetSubnetID(_ context.Context, chainID ids.ID) (ids.ID, error) {
-	if chainID == constants.PlatformChainID {
-		return constants.PrimaryNetworkID, nil
-	}
-
-	chainTx, _, err := vm.state.GetTx(chainID)
-	if err != nil {
-		return ids.Empty, fmt.Errorf(
-			"problem retrieving blockchain %q: %w",
-			chainID,
-			err,
-		)
-	}
-	chain, ok := chainTx.Unsigned.(*txs.CreateChainTx)
-	if !ok {
-		return ids.Empty, fmt.Errorf("%q is not a blockchain", chainID)
-	}
-	return chain.SubnetID, nil
-}
-
-// GetMinimumHeight returns the height of the most recent block beyond the
-// horizon of our recentlyAccepted window.
-//
-// Because the time between blocks is arbitrary, we're only guaranteed that
-// the window's configured TTL amount of time has passed once an element
-// expires from the window.
-//
-// To try to always return a block older than the window's TTL, we return the
-// parent of the oldest element in the window (as an expired element is always
-// guaranteed to be sufficiently stale). If we haven't expired an element yet
-// in the case of a process restart, we default to the lastAccepted block's
-// height which is likely (but not guaranteed) to also be older than the
-// window's configured TTL.
-//
-// If [UseCurrentHeight] is true, we will always return the last accepted block
-// height as the minimum. This is used to trigger the proposervm on recently
-// created subnets before [recentlyAcceptedWindowTTL].
-func (vm *VM) GetMinimumHeight(ctx context.Context) (uint64, error) {
-	if vm.Config.UseCurrentHeight {
-		return vm.GetCurrentHeight(ctx)
-	}
-
-	oldest, ok := vm.recentlyAccepted.Oldest()
-	if !ok {
-		return vm.GetCurrentHeight(ctx)
-	}
-
-	blk, err := vm.manager.GetBlock(oldest)
-	if err != nil {
-		return 0, err
-	}
-
-	// We subtract 1 from the height of [oldest] because we want the height of
-	// the last block accepted before the [recentlyAccepted] window.
-	//
-	// There is guaranteed to be a block accepted before this window because the
-	// first block added to [recentlyAccepted] window is >= height 1.
-	return blk.Height() - 1, nil
-}
-
-// GetCurrentHeight returns the height of the last accepted block
-func (vm *VM) GetCurrentHeight(context.Context) (uint64, error) {
-	lastAccepted, err := vm.manager.GetBlock(vm.state.GetLastAccepted())
-	if err != nil {
-		return 0, err
-	}
-	return lastAccepted.Height(), nil
 }
 
 func (vm *VM) CodecRegistry() codec.Registry {
